文件管理 重命名引发的血案

+++ app/src/main/java/com/stratagile/pnrouter/entity/file/File.kt
@@ -6,4 +6,4 @@ class FileOpreateType(var icon :String, var name : String) {
 
 }
 
-class UpLoadFile(var path : String, var isDownLoad : Boolean = true, var isComplete : Boolean = false, var isStop : Boolean = false, var progress : Int = 0, var total : Int = 0, var speed : Int = 0)
+class UpLoadFile(var path : String, var isDownLoad : Boolean = true, var isComplete : Boolean = false, var isStop : Boolean = false, var progress : Long = 0, var total : Long = 0, var speed : Int = 0)
